refactor: extract quota/rate-limit error helpers from standalone-chat
Move buildDailyLimitMessage, classifyQuotaError, buildRateLimitMessage,
parseRateLimitWaitSeconds, and PI_MAX_RATE_LIMIT_RETRIES into a new
lib/chat/quota-errors.ts. Pure code-move — no behavior change.

First step of breaking up the 9.8k-line standalone-chat.tsx. tsc passes,
all chat-store unit tests pass.
