+++ apps/screenpipe-app-tauri/components/standalone-chat.tsx
@@ -95,6 +95,13 @@ import { sanitizeToolCallXml } from "@/lib/utils/sanitize-tool-call-xml";
 import { useAutoSuggestions, type Suggestion } from "@/lib/hooks/use-auto-suggestions";
 import { SummaryCards, type ConnectionSetupSuggestion } from "@/components/chat/summary-cards";
 import { type CustomTemplate } from "@/lib/summary-templates";
+import {
+  buildDailyLimitMessage,
+  classifyQuotaError,
+  buildRateLimitMessage,
+  parseRateLimitWaitSeconds,
+  PI_MAX_RATE_LIMIT_RETRIES,
+} from "@/lib/chat/quota-errors";
 import { usePipes } from "@/lib/hooks/use-pipes";
 import { localFetch, getApiBaseUrl } from "@/lib/api";
 import { CONNECTIONS_UPDATED_EVENT } from "@/lib/connections-events";
@@ -553,83 +560,6 @@ const STATIC_MENTION_SUGGESTIONS: MentionSuggestion[] = [
  * Extract tier info from gateway error JSON embedded in error strings and
  * return a user-facing message appropriate to their actual subscription tier.
  */
-function buildDailyLimitMessage(errorStr: string): string {
-  try {
-    const isCostLimit = errorStr.includes("daily_cost_limit_exceeded");
-    const isRateLimit = errorStr.includes("rate limit") || errorStr.includes("Rate limit");
-
-    if (isRateLimit) {
-      return "This model is temporarily rate-limited. Try again in a few seconds, or switch to a different model.";
-    }
-
-    if (isCostLimit) {
-      // Don't leak the raw dollar cap — that's our internal margin. Frame it
-      // as an account-wide budget so the user understands why it fired even
-      // when they "didn't use much" (background pipes consume it too).
-      return "You've hit today's AI usage limit. This is an account-wide budget — background pipes count too. Switch to a free model (gemini-3-flash, haiku) or check Settings → Pipes for chatty schedules.";
-    }
-
-    const tierMatch = errorStr.match(/"tier":\s*"([^"]+)"/);
-    const tier = tierMatch?.[1];
-
-    if (tier === "subscribed") {
-      return "You've hit your daily limit. Switch to a free model (Qwen3 Coder, Gemini Flash) for unlimited usage.";
-    } else if (tier === "logged_in") {
-      return "You've used your free queries for today. Switch to a free model (Qwen3 Coder, Gemini Flash) for unlimited usage, or upgrade to Pro.";
-    } else {
-      return "You've used your free queries for today. Sign in for more, or switch to a free model (Qwen3 Coder, Gemini Flash).";
-    }
-  } catch {
-    return "You've reached your daily limit. Try a free model like Qwen3 Coder or Gemini Flash.";
-  }
-}
-
-function classifyQuotaError(errorStr: string): "daily" | "rate" | "none" {
-  const normalized = errorStr.toLowerCase();
-  const isDailyLimit =
-    normalized.includes("credits_exhausted") ||
-    normalized.includes("daily_limit_exceeded") ||
-    normalized.includes("daily_cost_limit_exceeded");
-  if (isDailyLimit) {
-    return "daily";
-  }
-
-  const isRateLimit =
-    normalized.includes("429") ||
-    normalized.includes("rate limit") ||
-    normalized.includes("rate_limit") ||
-    normalized.includes("requests per minute") ||
-    normalized.includes("too many requests");
-  return isRateLimit ? "rate" : "none";
-}
-
-function buildRateLimitMessage(errorStr: string): string {
-  const waitMatch = errorStr.match(/wait (\d+) seconds/i);
-  const waitTime = waitMatch ? waitMatch[1] : "a moment";
-  const isPerMinuteRate = /rate limit exceeded|requests per minute/i.test(errorStr);
-  return isPerMinuteRate
-    ? `Rate limited — please wait ${waitTime} seconds and try again.`
-    : "Rate limited — try again in a moment or switch to a different model.";
-}
-
-/** How many times a single turn auto-retries on a 429 before giving up. */
-const PI_MAX_RATE_LIMIT_RETRIES = 3;
-
-/**
- * Seconds to wait before retrying a rate-limited (429) request. Prefers the
- * gateway's structured `reset_in` hint, falls back to the "wait N seconds"
- * prose, then a safe default. Clamped to [1, 60].
- */
-function parseRateLimitWaitSeconds(errorStr: string): number {
-  const DEFAULT_WAIT = 10;
-  const resetMatch = errorStr.match(/"reset_in"\s*:\s*(\d+)/i);
-  const waitMatch = errorStr.match(/wait (\d+) seconds/i);
-  const raw = resetMatch?.[1] ?? waitMatch?.[1];
-  const secs = raw ? parseInt(raw, 10) : DEFAULT_WAIT;
-  if (!Number.isFinite(secs) || secs <= 0) return DEFAULT_WAIT;
-  return Math.min(Math.max(secs, 1), 60);
-}
-
 // Helper to get timezone offset string (e.g., "+1" or "-5")
 function getTimezoneOffsetString(): string {
   const offsetMinutes = new Date().getTimezoneOffset();
